requirements.txt
brotli==1.0.9
gmpy2==2.1.2
importlib-metadata==8.4.0
importlib-resources==6.4.0
ipykernel==6.29.5
ipywidgets==8.1.5
jaraco.text==3.12.1
mkl-fft==1.3.8
mkl-random==1.2.4
openai==1.42.0
ordered-set==4.1.0
pickleshare==0.7.5
pinecone==5.0.1
pip-chill==1.0.3
pysocks==1.7.1
python-dotenv==1.0.1
tomli==2.0.1
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.44.2
win-inet-pton==1.1.0
